pyproject.toml
[project]
name = "acs-pums-pipeline"
version = "0.1.0"
description = "Ingestion pipeline for ACS and PUMS census data → Parquet"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31",
    "pandas>=2.2,<3",
    "pyarrow>=15.0",
    "duckdb>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
notebooks = [
    "jupyterlab>=4.0",
    "matplotlib>=3.8",
]

[tool.setuptools]
packages = ["pipeline"]
